requirements.txt
streamlit>=1.40.0
google-genai>=1.0.0
Pillow>=10.0.0
pillow-heif>=0.18.0
python-dotenv>=1.0.0
